# internal/handler/authenticationHandler.go
package handler

import (
	"encoding/json"
	"medivault-service/internal/services"
	"medivault-service/pkg/utils"
	"net/http"
)

type AuthenticationHandler struct {
	Service services.UserService
}

func NewAuthenticationHandler(service services.UserService) *AuthenticationHandler {
	return &AuthenticationHandler{service}
}

// Login logs in a user
//
//	@Summary		Login a user
//	@Description	Logs in a user
//	@Tags			authentication
//	@Accept			json
//	@Produce		json
//	@Param			email		body		string	true	"User email"
//	@Param			password	body		string	true	"User password"
//	@Success		200			{object}	User
//	@Failure		400			{object}	utils.ErrorResponse
//	@Failure		500			{object}	utils.ErrorResponse
//	@Router			/login [post]
func (auth *AuthenticationHandler) Login(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented, // 501
		Message: "Login functionality is not implemented.",
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}

// Logout logs out a user
//
//	@Summary		Logout a user
//	@Description	Logs out a user
//	@Tags			authentication
//	@Accept			json
//	@Produce		json
//	@Success		200	{object}	User
//	@Failure		400	{object}	utils.ErrorResponse
//	@Failure		500	{object}	utils.ErrorResponse
//	@Router			/logout [post]
func (auth *AuthenticationHandler) Logout(w http.ResponseWriter, r *http.Request) {
	// Create the error response
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented, // 501
		Message: "Logout functionality is not implemented.",
	}

	// Set the response header content type to application/json
	w.Header().Set("Content-Type", "application/json")
	// Write the status code
	w.WriteHeader(http.StatusNotImplemented)
	// Encode the error response to JSON and write it to the response writer
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		// Handle potential encoding errors
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}

// RefreshToken refreshes a user token
//
//	@Summary		Refresh a user token
//	@Description	Refreshes a user token
//	@Tags			authentication
//	@Accept			json
//	@Produce		json
//	@Success		200	{object}	User
//	@Failure		400	{object}	utils.ErrorResponse
//	@Failure		500	{object}	utils.ErrorResponse
//	@Router			/refresh [post]
func (auth *AuthenticationHandler) RefreshToken(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented, // 501
		Message: "RefreshToken functionality is not implemented.",
	}

	// Set the response header content type to application/json
	w.Header().Set("Content-Type", "application/json")
	// Write the status code
	w.WriteHeader(http.StatusNotImplemented)
	// Encode the error response to JSON and write it to the response writer
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		// Handle potential encoding errors
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}