# internal/handler/patientFileCollectionHandler.go
package handler

import (
	"encoding/json"
	"medivault-service/internal/services"
	"medivault-service/pkg/utils"
	"net/http"
)

type PatientFileCollectionHandler struct {
	Service services.PatientFileCollectionService
}

// GET all patientFileCollection by patientID
//
//	@Summary		Get all patientFileCollections by patient ID
//	@Description	Retrieve all patientFileCollections by patient ID
//	@Tags			patientFileCollections
//	@Accept			json
//	@Produce		json
//	@Param			id	path		int	true	"Patient ID"
//	@Success		200	{object}	PatientFileCollection
//	@Failure		400	{object}	utils.ErrorResponse
//	@Failure		404	{object}	utils.ErrorResponse
//	@Failure		500	{object}	utils.ErrorResponse
//	@Router			/patientFileCollections/{id} [get]
func (pfcH *PatientFileCollectionHandler) GetAllPatientFileCollectionsByPatientID(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented, // 501
		Message: "GetAllPatientFileCollectionsByPatientID functionality is not implemented.",
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}

// Create patientFileCollection
//
//	@Summary		Create a new patient file collection
//	@Description	Register a new patient file collection in the system
//	@Tags			patientFileCollections
//	@Accept			json
//	@Produce		json
//	@Param			patientFileCollection	body		PatientFileCollection	true	"PatientFileCollection data"
//	@Success		201						{object}	PatientFileCollection
//	@Failure		400						{object}	ErrorResponse
//	@Failure		500						{object}	ErrorResponse
//	@Router			/patientFileCollections [post]
func (pfcH *PatientFileCollectionHandler) CreatePatientFileCollection(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented, // 501
		Message: "CreatePatientFileCollection functionality is not implemented.",
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}

// Update patientFileCollection
//
//	@Summary		Update a patient file collection
//	@Description	Update a patient file collection in the system
//	@Tags			patientFileCollections
//	@Accept			json
//	@Produce		json
//
//	@Param			id	path		int	true	"PatientFileCollection ID"
//
//	@Success		200	{object}	PatientFileCollection
//	@Failure		400	{object}	ErrorResponse
//	@Failure		404	{object}	ErrorResponse
//	@Failure		500	{object}	ErrorResponse
//	@Router			/patientFileCollections/{id} [put]
func (pfcH *PatientFileCollectionHandler) UpdatePatientFileCollection(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented, // 501
		Message: "UpdatePatientFileCollection functionality is not implemented.",
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}

// Delete patientFileCollection
func (pfcH *PatientFileCollectionHandler) DeletePatientFileCollection(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented,
		Message: "DeletePatientFileCollection functionality is not implemented.",
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {

		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}

// Add patientFile to patientFileCollection
func (pfcH *PatientFileCollectionHandler) AddPatientFileToPatientFileCollection(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented,
		Message: "AddPatientFileToPatientFileCollection functionality is not implemented.",
	}

	w.Header().Set("Content-Type", "application/json")

	w.WriteHeader(http.StatusNotImplemented)

	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}

// Remove patientFile from patientFileCollection
func (pfcH *PatientFileCollectionHandler) RemovePatientFileFromPatientFileCollection(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented,
		Message: "RemovePatientFileFromPatientFileCollection functionality is not implemented.",
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}

// Get patientFile from patientFileCollection
func (pfcH *PatientFileCollectionHandler) GetPatientFileFromPatientFileCollection(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented,
		Message: "GetPatientFileFromPatientFileCollection functionality is not implemented.",
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}