# internal/handler/patientFileHandler.go
package handler

import (
	"encoding/json"
	"medivault-service/internal/services"
	"medivault-service/pkg/utils"
	"net/http"
)

type PatientFileHandler struct {
	Service services.PatientFileService
}

// CreatePatientFile creates a new patient file
//
//	@Summary		Create a new patient file
//	@Description	Register a new patient file in the system
//	@Tags			patientFiles
//	@Accept			json
//	@Produce		json
//	@Param			patientFile	body		PatientFile	true	"PatientFile data"
//	@Success		201			{object}	PatientFile
//	@Failure		400			{object}	utils.ErrorResponse
//	@Failure		500			{object}	utils.ErrorResponse
//	@Router			/patientFiles [post]
func (pfH *PatientFileHandler) CreatePatientFile(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented, // 501
		Message: "CreatePatientFile functionality is not implemented.",
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}

// GetAllPatientFilesByPatientID
//
//	@Summary		Get all patient files by patient ID
//	@Description	Retrieve all patient files by patient ID
//	@Tags			patientFiles
//	@Accept			json
//	@Produce		json
//	@Param			id	path		int	true	"Patient ID"
//	@Success		200	{object}	PatientFile
//	@Failure		400	{object}	utils.ErrorResponse
//	@Failure		404	{object}	utils.ErrorResponse
//	@Failure		500	{object}	utils.ErrorResponse
//	@Router			/patientFiles/{id} [get]
func (pfH *PatientFileHandler) GetAllPatientFilesByPatientID(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented, // 501
		Message: "GetAllPatientFilesByPatientID functionality is not implemented.",
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}

// UpdatePatientFile updates a patient file
//
//	@Summary		Update a patient file
//	@Description	Update a patient file in the system
//	@Tags			patientFiles
//	@Accept			json
//	@Produce		json
//	@Param			id			path		int			true	"PatientFile ID"
//	@Param			patientFile	body		PatientFile	true	"PatientFile data"
//	@Success		200			{object}	PatientFile
//	@Failure		400			{object}	utils.ErrorResponse
//	@Failure		404			{object}	utils.ErrorResponse
//	@Failure		500			{object}	utils.ErrorResponse
//	@Router			/patientFiles/{id} [put]
func (pfH *PatientFileHandler) UpdatePatientFile(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented, // 501
		Message: "UpdatePatientFile functionality is not implemented.",
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}

// DeletePatientFile deletes a patient file
//
//	@Summary		Delete a patient file
//	@Description	Delete a patient file in the system
//	@Tags			patientFiles
//	@Accept			json
//	@Produce		json
//	@Param			id	path		int	true	"PatientFile ID"
//	@Success		204	{object}	PatientFile
//	@Failure		400	{object}	utils.ErrorResponse
//	@Failure		404	{object}	utils.ErrorResponse
//	@Failure		500	{object}	utils.ErrorResponse
//	@Router			/patientFiles/{id} [delete]
func (pfH *PatientFileHandler) DeletePatientFile(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented, // 501
		Message: "DeletePatienFile functionality is not implemented.",
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}