# internal/handler/types.go
package handler

import "medivault-service/internal/db/generated"

type Role = *generated.Role

type PatientFile = *generated.PatientFile

type User = *generated.User

type PatientFileCollection = *generated.PatientCollection